% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedFig.R
\name{embedFig}
\alias{embedFig}
\title{Generate a figure bitmap file with embedded metadata}
\usage{
embedFig(plot.fun, arg.list, full.metadata, format = "png", filename, ...)
}
\arguments{
\item{plot.fun}{Character. Name of the plotting function}

\item{arg.list}{List of arguments passed to the plotting function. The list is functional, and not 
merely descriptive, so all arguments (including required input grids) need to be passed here.}

\item{full.metadata}{An igraph-class containing the RDF graph with the full metadata structure.}

\item{format}{Format of the bitmap file. Currently only the \code{"png"} option is available}

\item{filename}{The name of the output file.}

\item{...}{further optional arguments passed to \code{\link{png}} (width, height etc.)}
}
\value{
A temporary file of the specified format
}
\description{
The function calls a plotting function, and writes to a user-defined bitmap file, including
the full metadata description provided by the user. This information is a compressed JSON representation.
}
\details{
The function performs the following steps:
\enumerate{
\item Write the output figure in a temporary file
\item Serialize the RDF graph into json and write it into a temporary file
\item Read the temporary json file and compress (gzip)
\item Read the temporary image file, and re-write into the destination file (as given in \code{filename}),
appending the compressed metadata
}
}
\seealso{
Other graphical.outputs: \code{\link{graph2json}}
}
\author{
J. Bedia
}
