% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedFig.R
\name{embedJSON}
\alias{embedJSON}
\title{Embed JSON-LD}
\usage{
embedJSON(png.in, json.file, png.out)
}
\arguments{
\item{png.in}{Full path to a target png file}

\item{json.file}{File path to JSON-LD}

\item{png.out}{Output png file path}
}
\value{
A png with embeded METACLIP provenance
}
\description{
Embeds a JSON file into a png as METACLIP metadata
}
\seealso{
Other graphical.outputs: \code{\link{embedFig}},
  \code{\link{graph2json}}
}
\author{
J Bedia
}
\concept{graphical.outputs}
