% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip.graph.Command.R
\name{formatCommandCallString}
\alias{formatCommandCallString}
\title{Format literal command calls}
\usage{
formatCommandCallString(text)
}
\arguments{
\item{text}{A literal command call given as a character string}
}
\value{
A literal command call, without blank spaces or line returns
}
\description{
Prepare character strings to be adequately displayed by the METACLIP Interpreter
}
\author{
J Bedia
}
\keyword{internal}
