% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProperty-helpers.R
\name{getHasProbabilityThreshold}
\alias{getHasProbabilityThreshold}
\title{Get the hasProbabilityValue property}
\usage{
getHasProbabilityThreshold(ForecastRepresentation, arg.list)
}
\arguments{
\item{ForecastRepresentation}{Character string designating the ForecastRepresentation Class}

\item{arg.list}{Argument list passed to \code{veriApply}}
}
\value{
Either \code{NULL} (if the property does not apply), or a list of attributes to be appended to the node.
}
\description{
Infer the hasProbabilityValue property from the arguments passed to \code{veriApply}
}
\seealso{
Other getProperty.helpers: \code{\link{getHasAbsoluteThreshold}},
  \code{\link{getHasTimeStep}}
}
\author{
J Bedia
}
\keyword{internal}
