% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProperty-helpers.R
\name{getHasTimeStep}
\alias{getHasTimeStep}
\title{Get ISO-8601 time interval definition}
\usage{
getHasTimeStep(grid)
}
\arguments{
\item{grid}{A climate4R grid}
}
\value{
A character string with the time resolution specification compliant with ISO-8601 (or \code{"unknown"})
}
\description{
Maps the time resolution output of \code{\link[transformeR]{getTimeResolution}} into a ISO-8601 duration specification
}
\references{
\url{https://en.wikipedia.org/wiki/ISO_8601#Durations}
}
\seealso{
Other getProperty.helpers: \code{\link{getHasAbsoluteThreshold}},
  \code{\link{getHasProbabilityThreshold}}
}
\author{
J Bedia
}
\keyword{internal}
