% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-helpers.R
\name{getNodeAttributes}
\alias{getNodeAttributes}
\title{Return all node attributes}
\usage{
getNodeAttributes(graph, vertex.index = NULL, vertex.name = NULL)
}
\arguments{
\item{graph}{\code{\link[igraph]{igraph}} graph}

\item{vertex.index}{Alternatively, instead of \code{vertex.name}, an integer
indicating the index position of the vertex in the graph.}

\item{vertex.name}{Character string, with the target vertex name}
}
\value{
A list of (non-empty) attributes belonging to the vertex
}
\description{
Return a list of all (non-empty) node attributes
}
\seealso{
\code{\link{getNodeIndexbyName}}

Other graph.helpers: \code{\link{getNodeIndexbyName}},
  \code{\link{getPlainPropertyList}}
}
\author{
J Bedia
}
