% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-helpers.R
\name{getNodeIndexbyName}
\alias{getNodeIndexbyName}
\title{Get node indices}
\usage{
getNodeIndexbyName(graph, names)
}
\arguments{
\item{graph}{\code{\link[igraph]{igraph}} graph}

\item{names}{Character vector of names of the nodes whose index is queried}
}
\value{
A vector of index positions corresponding to the input names
}
\description{
Get the node index positions from their names
}
\examples{
require(igraph)
graph <- make_graph(c("A", "B", "B", "C", "C", "D", "Pepe"), directed = FALSE)
getNodeIndexbyName(graph, "Pepe")
getNodeIndexbyName(graph, c("A","D","Pepe"))
}
\seealso{
Other graph.helpers: \code{\link{getNodeAttributes}},
  \code{\link{getPlainPropertyList}}
}
\author{
J Bedia
}
