% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-helpers.R
\name{getPlainPropertyList}
\alias{getPlainPropertyList}
\title{Return most node attributes}
\usage{
getPlainPropertyList(g, node.index)
}
\arguments{
\item{g}{\code{\link[igraph]{igraph}} graph}

\item{node.index}{An integer indicating the index position of the vertex in the graph.}
}
\value{
A list of (non-empty) attributes belonging to the vertex (lists are vectorized)
}
\description{
Return almost all node attributes, but the name, classname, etc...
}
\seealso{
Other graph.helpers: \code{\link{getNodeAttributes}},
  \code{\link{getNodeIndexbyName}}
}
\author{
J Bedia
}
\keyword{internal}
