% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClass-helpers.R
\name{getQualityAspect}
\alias{getQualityAspect}
\title{Get QualityAspect class value}
\source{
\url{https://docs.google.com/presentation/d/1-H-d5X2kdA6-0nqnCd5yLs_V1DeTzuTGw2g3bLhtj7I/present#slide=id.g23b883c69b_0_194}
}
\usage{
getQualityAspect(verifun)
}
\arguments{
\item{arg.list}{Argument list passed to \code{veriApply}}
}
\value{
A character string of the QualityAspect Class
}
\description{
Infer the Class inheriting from Abstract Class QualityAspect from the value of verifun argument
}
\seealso{
Other getClass.helpers: \code{\link{getForecastRepresentation}}
}
\author{
J Bedia
}
\keyword{internal}
