% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClass-helpers.R
\name{getVerificationLabel}
\alias{getVerificationLabel}
\title{Get Verification class node label}
\usage{
getVerificationLabel(verifun)
}
\arguments{
\item{arg.list}{Argument list passed to \code{veriApply}}
}
\value{
A character string of the verification measure label
}
\description{
Give a label to the specific measure used for the Verification node, as inferred from the value of verifun argument
}
\author{
J Bedia
}
\keyword{internal}
