% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2json.R
\name{graph2json}
\alias{graph2json}
\title{RDF graph serialization to JSON-LD}
\usage{
graph2json(graph, output.file)
}
\arguments{
\item{graph}{An i-graph class graph}

\item{output.file}{Character string. Output path}
}
\value{
A JSON-LD representation of the metadata structure
}
\description{
Takes an igraph-class RDF graph and write it in JSON-LD
}
\references{
A useful app to check/test JSON-LD mark <https://json-ld.org/playground/>
}
\seealso{
Other graphical.outputs: \code{\link{embedFig}}
}
\author{
J Bedia
}
