% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2json.R
\name{isIndividualInstance}
\alias{isIndividualInstance}
\title{Individual identification}
\usage{
isIndividualInstance(node.name)
}
\value{
A Logical flag
}
\description{
Is the node representing and individual instance, or a generic class entity?
}
\author{
J Bedia
}
\keyword{internal}
