% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{knownClassIndividuals}
\alias{knownClassIndividuals}
\title{List all individuals from a class}
\usage{
knownClassIndividuals(classname, vocabulary = "datasource")
}
\arguments{
\item{classname}{The parent class from which the individual instances are queried}

\item{vocabulary}{The target vocabulary name. Possible values are \code{"datasource"} (the default),
\code{"calibration"}, \code{"verification"} and \code{"graphical_output"}.}
}
\value{
A character vector of individuals for that class
}
\description{
List defined individuals pertaining to a specific class from a METACLIP vocabulary
}
\details{
The function will check the existing individuals in the latest stable datasource 
ontology release, available in \url{http://metaclip.predictia.es/datasource/datasource.owl}
}
\note{
The function won't work if there is no internet connection, or any other connection problem prevents
 read access to the ontology file.
}
\examples{
knownClassIndividuals("ModellingCenter")
knownClassIndividuals("ETCCDI")
# In case a class is misspelled or it has no individuals, 
# an empty vector is returned, with a warning:
knownClassIndividuals("ETCDDI")
knownClassIndividuals("Dataset")
}
\author{
J Bedia, D. San-Martín
}
