% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{knownPackageVersions}
\alias{knownPackageVersions}
\title{List of valid package versions for metaclipR}
\usage{
knownPackageVersions()
}
\description{
Print a table with the valid version range of known packages. For the listed packages and version ranges, 
metaclipR fuctions are guaranteed to correctly map arguments/functions to the corresponding METACLIP classes/properties.
}
