% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.AnomalyCalculation.R
\name{metaclipR.Anomaly}
\alias{metaclipR.Anomaly}
\title{Directed metadata graph construction for Anomaly transformations}
\usage{
metaclipR.Anomaly(graph, package = "transformeR", version = "1.4.4",
  fun = "scaleGrid", time.frame = "seasonal", arg.list = NULL,
  referenceGraph = NULL, clim.cell.method = "mean",
  disable.command = FALSE)
}
\arguments{
\item{graph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing
the input grid whose anomaly is to be computed, plus the terminal node from which the Anomaly Step will hang}

\item{package}{package}

\item{version}{version}

\item{fun}{function name. Unused (set to \code{"scaleGrid"})}

\item{time.frame}{Time frame considered for climatological reference calculation}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{referenceGraph}{An output from a previous \pkg{metaclipR} function containing a list with the i-graph class object containing the reference Transformation-class object
used as base to compute the climatology, plus the name of its terminal node}

\item{clim.cell.method}{Optional. cell method for climatology calculation. Default to \code{"mean"}.}

\item{disable.command}{Better not to touch. For internal usage only (used to re-use most of the code 
in other functions, but skipping command tracking)}
}
\description{
Build a directed metadata graph describing an anomaly Transformation
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
is a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

