% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.BiasCorrection.R
\name{metaclipR.BiasCorrection}
\alias{metaclipR.BiasCorrection}
\title{Directed metadata graph construction for bias correction steps}
\usage{
metaclipR.BiasCorrection(package = "downscaleR", version = "3.0.0",
  fun = "biasCorrection", arg.list, graph, TrainingGraph, ReferenceGraph,
  BC.method, BC.class = "BiasCorrection", isDefinedBy = NULL,
  comment = NULL, hasProbCharacter = NULL, isMultivariable = NULL,
  isMultisite = NULL)
}
\arguments{
\item{package}{bias correction package. Default to \code{"downscaleR"}}

\item{version}{A character string of the package version (e.g. as returned by \code{\link[utils]{packageVersion}})}

\item{fun}{Validation function. Default to \code{"biasCorrection"}.}

\item{arg.list}{argument list passed by \code{fun}.}

\item{graph}{metaclipR output containing the data top be bias-corrected.}

\item{TrainingGraph}{metaclipR output containing the training data (e.g. 20C3M/historical scenario in
climate change applications etc.)}

\item{ReferenceGraph}{metaclipR output containing the reference predictand (typically observations)}

\item{BC.method}{Character string indicating the name of the bias correction method.
For a list of known (instantiable) BiasCorrection methods see \code{knownClassIndividuals("BiasCorrection", "calibration")}}

\item{BC.class}{Character string indicating the specific subclass of the Calibration superclass being asserted.
Accepted values are defined in the Calibration ontology (see Details). By default, the superclass \code{"BiasCorrection"}
is assigned, but other more specific subclasses should be indicated when relevant.}

\item{isDefinedBy}{An optional character string indicating where or by who the method has been defined 
(e.g. paper publication etc.). This is encoded as rdfs:isDefinedBy data property associated to the calibration method).}

\item{comment}{An optional character string with a comment on the method (encoded as rdfs:comment data property
associated to the calibration method)}

\item{hasProbCharacter}{Default to NULL (unspecified). It referes to the probabilistic nature of the method. 
Two possible disjoint values: \code{"deterministic"} or \code{"stochastic"}}

\item{isMultivariable}{Default to NULL (unspecified). isMultisite Default to NULL (unspecified), otherwise logical. Is the bias correction method envisaged to 
deal with multisite predictions? See References.}

\item{isMultisite}{Default to NULL (unspecified), otherwise logical. Is the bias correction method envisaged to 
deal with multisite predictions? See References.}
}
\description{
Build a directed metadata graph from bias correction routines
}
\details{
This function takes as reference the semantics defined in the Calibration ontology defined in the 
Metaclip Framework (\url{http://metaclip.predictia.es/}). These in turn are partially based on the VALUE Framewrok (Gutiérrez et al. 2018)
}
\references{
Gutiérrez et al, 2018. An intercomparison of a large ensemble of statistical downscaling methods over Europe: 
Results from the VALUE perfect predictor cross-validation experiment. International Journal of Climatology. 
https://doi.org/10.1002/joc.5462
}
\author{
D. San Martín, J. Bedia
}
