% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.etccdi.R
\name{metaclipR.ClimateIndex}
\alias{metaclipR.ClimateIndex}
\title{METACLIP description of climate index calculation steps}
\usage{
metaclipR.ClimateIndex(
  graph,
  package = "climate4R.climdex",
  version = as.character(packageVersion(package)),
  output = NULL,
  index.code,
  fun = "climdexGrid",
  arg.list = NULL,
  new.time.res = NULL,
  time.step = NULL,
  cell.method = NULL,
  disable.command = FALSE
)
}
\arguments{
\item{graph}{Output from previous metaclipR function. A list with an i-graph class object plus the name of the parent node 
from which the climate index step hangs. \code{climdexGrid} from package \pkg{climate4R.climdex} is indicated in argument \code{fun}.}

\item{package}{package}

\item{version}{version}

\item{output}{Optional. The output climate4R object name, as character string}

\item{index.code}{Character string, indicating the specific code of the index according to the ETCCDI definitions (see Details)}

\item{fun}{function name. Unused (set to \code{"climdexGrid"})}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{new.time.res}{Character string. Recommended an ISO8601 representation of temporal resolution, 
as a "duration". Typical values are "P1H" (hourly), "P3H" (3-hourly), "P1D" (daily), "P1M" (monthly) etc.
If omitted, the last temporal resolution given is assumed to be preserved, and this is not updated.}

\item{time.step}{Optional. Time step data property}

\item{cell.method}{Optional. Cell method. Default to the \code{index.code} value.}

\item{disable.command}{Better not to touch. For internal usage only (used to re-use most of
the code in other functions, but skipping command tracking)}
}
\description{
METACLIP description of climate index calculation steps
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
is a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

\references{
For reference on ISO8601 definition \url{https://en.wikipedia.org/wiki/ISO_8601#Durations}
}
