% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Climatology.R
\name{metaclipR.Climatology}
\alias{metaclipR.Climatology}
\title{Directed metadata graph construction for Climatology transformations}
\usage{
metaclipR.Climatology(
  package = "transformeR",
  version = as.character(packageVersion(package)),
  graph,
  fun = "climatology",
  arg.list = NULL,
  disable.command = FALSE,
  dc.description = NULL
)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{graph}{A previous metaclipR data structure from which the current step follows}

\item{fun}{function name. Unused (set to \code{"climatology"})}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{disable.command}{Better not to touch. For internal usage only (used to re-use most of the code in other
functions, but skipping command tracking)}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:Climatology-class node.}
}
\value{
A named list with the updated graph in element \code{"graph"} and the parent node name (\code{"parentnodename"}),
needed for linking subsequent operations.
}
\description{
Build a directed metadata graph describing a Transformation on a 
climate4R grid that is a Climatology
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
is a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

\references{
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other transformation: 
\code{\link{metaclip.graph.Command}()},
\code{\link{metaclipR.Aggregation}()},
\code{\link{metaclipR.AnomalyCalculation}()},
\code{\link{metaclipR.Binding}()},
\code{\link{metaclipR.Dataset}()},
\code{\link{metaclipR.Ensemble}()},
\code{\link{metaclipR.Ensemble2}()},
\code{\link{metaclipR.Interpolation}()},
\code{\link{metaclipR.etccdi}()}
}
\author{
D. San Martín, J. Bedia
}
\concept{transformation}
