% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Dataset.R
\name{metaclipR.Dataset}
\alias{metaclipR.Dataset}
\title{Directed metadata graph construction for Data Sources}
\usage{
metaclipR.Dataset(
  Dataset.name = NULL,
  Dataset.subclass = NULL,
  DataProvider = NULL,
  ModellingCenter = NULL,
  DataProvider.URL = NULL,
  Project = NULL,
  RCM = NULL,
  GCM = NULL,
  Run = NULL
)
}
\arguments{
\item{Dataset.name}{Name (label) of the Dataset. There is a number of already known Datasets, pertaining to the User Data Gateway 
Public datasets of the climate4R Framework (see References). Type \code{showUDGDatasources()} for details.
 If the argument corresponds to any of these named instances, the associated provenance information will be automatically recorded,
 and all other arguments can be omitted.}

\item{Dataset.subclass}{A character string indicating the Dataset subclass. The only (and disjoint) possible values 
are \code{"MultiDecadalSimulation"}, \code{"ObservationalDataset"}, \code{"Reanalysis"}, \code{"SeasonalHindcast"},
 \code{"SeasonalOperationalForecast"} and \code{"ShortRangeForecast"}.}

\item{DataProvider}{A Character string indicating the data provider of the Dataset (e.g. \code{"UDG"}, 
for datasets accessed through the Santander MetGroup User Data Gateway, see References).
 NOTE: Any ds:ModellingCenter can be also be ds:DataProvider. For a list of known (instantiable) data providers see
  \code{knownClassIndividuals("DataProvider")} and \code{knownClassIndividuals("ModellingCenter")}.}

\item{ModellingCenter}{Optional character vector, e.g. \code{"ECMWF"}, or \code{c("SMHI","KNMI")}, in case there are two or more modelling centers involved.
See \code{knownClassIndividuals("ModellingCenter")} for a list of known (instantiable) modelling centers.}

\item{DataProvider.URL}{Optional: The URL from which the data was accessed.}

\item{Project}{Optional: Character string indicating the Project leading to the generatrion of the Dataset (e.g. \code{"CORDEX"}).
See \code{knownClassIndividuals("Project")} for a list of known (instantiable) modelling centers.}

\item{RCM}{Optional. Character string (length one). Simulation RCM model prducing the dataset (e.g. \code{"RACMO22E"}). 
See \code{knownClassIndividuals("RCM")} for a list of known (instantiable) RCMs.}

\item{GCM}{Optional. Character string (length one). GCM producing the dataset
See \code{knownClassIndividuals("GCM")} for a list of known (instantiable) GCMs (e.g. \code{"EC-EARTH"}).}

\item{Run}{Optional. When relevant, a character string (length one) indicating the model run (e.g. \code{"r1i1p1"})}
}
\description{
Build a directed metadata graph describing a data source. This is usually the initial 
step to build METACLIP graphs
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org/}).
 Many of the classes represented have individuals defined that can be directly instantiated instead of the Class itself. This has the advantage
 of providing further annotations relevant for provenance description (e.g. references, relevant URLs, comments etc.).
  This is automatically handled by the function when there is perfect match between the argument value and the known individual name.
   To ensure consistency, the use of the helper function \code{\link{knownClassIndividuals}} is recommended.
}
\examples{

# showUDGDatasources() will display all the known datasets from the User Data Gateway.
# With just indicating the dataset name (and the provider, which is not assumed to be UDG,
# so dataset metadata can be reused if accessed through other providers), the rest of metadata
# will be appennded. Further annotations are included inside the nodes (e.g. access URL,
# model run etc. Not shown by the i-graph plot display)

# A GCM 
dataset <- "CMIP5_CNRM-CERFACS-CNRM-CM5_historical"
gcm.dataset <- metaclipR.Dataset(Dataset.name = dataset, DataProvider = "UDG")
plot(gcm.dataset$graph)                              

# An RCM (note the hadDrivingGCM property)
dataset <- "EUROCORDEX11_MOHC-HadGEM2-ES_r1i1p1_RCA4_v1_rcp45"
rcm.dataset <- metaclipR.Dataset(Dataset.name = dataset,
                                 DataProvider = "UDG")
plot(rcm.dataset$graph)                              

# When dealing with known Individuals, the node is kept fix whenever it is necessary to refer to it:
metadata <- metaclipR.Dataset(Dataset.name = "E-OBS_v14_0.25regular",
                              DataProvider = "KNMI")
plot(metadata$graph)
}
\seealso{
Other transformation: 
\code{\link{metaclip.graph.Command}()},
\code{\link{metaclipR.Aggregation}()},
\code{\link{metaclipR.AnomalyCalculation}()},
\code{\link{metaclipR.Binding}()},
\code{\link{metaclipR.Climatology}()},
\code{\link{metaclipR.Ensemble}()},
\code{\link{metaclipR.Ensemble2}()},
\code{\link{metaclipR.Interpolation}()},
\code{\link{metaclipR.etccdi}()}
}
\author{
D. San Martín, J. Bedia
}
\concept{transformation}
