% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.DatasetSubset.R
\name{metaclipR.DatasetSubset}
\alias{metaclipR.DatasetSubset}
\title{Directed metadata graph construction for DatasetSubsets}
\usage{
metaclipR.DatasetSubset(package = "transformeR", version = "1.3.2",
  graph = NULL, fun = NULL, arg.list = NULL, output = NULL,
  disable.command = FALSE)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{graph}{A previously initialized metaclipR output list ($graph + $parentnodename). Default to \code{NULL}, which starts a new empty graph}

\item{fun}{function name (default, and only value accepted is \code{"subsetGrid"})}

\item{arg.list}{arg.list. See details}

\item{output}{A climate4R grid, resulting from the application of \code{\link[transformeR]{subsetGrid}}.}

\item{disable.command}{Better not to touch. For internal usage only (used to re-use most of the code in other functions, but skipping command tracking)}
}
\value{
A named list with the updated graph in element \code{"graph"} and the parent node name,
sometimes needed for linking subsequent operations.

\strong{Argument list}

A list of arguments passed to \code{\link{subsetGrid}} is required. The different 
arguments are explained in the the help menu of \code{subsetGrid}.
The following list need to be specified. The default values are also indicated:

\itemize{
\item \code{var = NULL}
\item \code{lonLim = NULL}
\item \code{latLim = NULL} 
\item \code{members = NULL} 
\item \code{years = NULL} 
\item \code{outside = FALSE}
\item \code{drop = TRUE}
}
}
\description{
Build a directed metadata graph from climate4R grids that are subsets of Datasets
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\examples{
data(S4_tas_iberia)
require(transformeR)
require(igraph)
# An example subset from S4_tas_iberia
arg.list <- list(grid = S4_tas_iberia,
                 members = 1:3, # First three members (out of 15)
                 season = NULL,
                 years = 1991:1995, # Period 1991-1995
                 lonLim = c(-5,3), # longitudinal extent
                 latLim = c(37,43),
                 drop = FALSE) # latitudinal extent
# Dataset subset is applied using subsetGrid:
dataset.subset <- do.call("subsetGrid", args = arg.list)
# Just for illustration, visualization of the climatology:
plotClimatology(climatology(dataset.subset, by.member = FALSE),
                 backdrop.theme = "countries")
# Encoding provenance: 
dsgraph <- metaclipR.DatasetSubset(package = "transformeR",
                                   version = "1.1.1",
                                   fun = "subsetGrid",
                                   arg.list = arg.list,
                                   output = "dataset.subset")
# The Graph containing the metadata has been created.
# This is a bit congested, but note that it is not conceived to be visualized in R
plot(dsgraph$graph, vertex.size = 5, edge.arrow.size=0.1)
}
\references{
\href{https://docs.google.com/presentation/d/1CQyxVIj501N7VylMR9i_T_XwFDId6MDNvpPnuaLXgnI/present}{Visual schema of the data transformation ontology}

\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other subsetting: \code{\link{metaclipR.loadeR}}
}
\author{
D. San Martín, J. Bedia
}
