% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Ensemble2.R
\name{metaclipR.Ensemble2}
\alias{metaclipR.Ensemble2}
\title{Construct a METACLIP representation of the ds:Ensemble class and ds:GrandEnsemble subclass}
\usage{
metaclipR.Ensemble2(
  graph.list,
  grand.ensemble = FALSE,
  named.individual = NULL,
  dc.description = NULL
)
}
\arguments{
\item{graph.list}{A \code{metaclipR} object data list, each element being the
graph defining each ensemble member}

\item{grand.ensemble}{Logical. If \code{TRUE}, the resulting ds:Ensemble will be annotated as its sub-class ds:GrandEnsemble.}

\item{named.individual}{Character string. The resulting ds:Ensemble (or its subclass ds:GrandEnsemble if \code{grand.ensemble} is set to \code{TRUE})
will be annotated as a named individual (i.e. with a "ds:name" annotation). Default to \code{NULL} and not used.}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:Ensemble-class node.}
}
\description{
Build a directed metadata graph describing an Ensemble transformation from two or more
ds:Steps
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\note{
This function is a variant of \code{\link{metaclipR.Ensemble}} that allows to create both ds:Ensemble 
and ds:GrandEnsemble instances, and to annotate them as named individuals.
The original function is still available for backward compatibility,
but it is recommended to use this new one instead.
}
\seealso{
Other transformation: 
\code{\link{metaclip.graph.Command}()},
\code{\link{metaclipR.Aggregation}()},
\code{\link{metaclipR.AnomalyCalculation}()},
\code{\link{metaclipR.Binding}()},
\code{\link{metaclipR.Climatology}()},
\code{\link{metaclipR.Dataset}()},
\code{\link{metaclipR.Ensemble}()},
\code{\link{metaclipR.Interpolation}()},
\code{\link{metaclipR.etccdi}()}
}
\author{
J. Bedia
}
\concept{transformation}
