% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.EnsoPlume.R
\name{metaclipR.EnsoPlume}
\alias{metaclipR.EnsoPlume}
\title{Directed metadata graph construction for ENSO plume graphical products}
\usage{
metaclipR.EnsoPlume(package = "visualizeR", version = "1.1.1",
  fun = "ensoPlumeC3S", multimodel = FALSE, input.obs.grid = NULL,
  input.fcst.grid = NULL, arg.list = NULL, FcstMetadata,
  ObsMetadata = NULL, VerificationMetadata = NULL)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{fun}{function name.}

\item{multimodel}{Logical. Is the forecast done with a multimodel ensemble dataset?. Default
to \code{FALSE} (This is needed to ascertain the subtypes of ForecastSystemUncertainty communicated
by the product, i.e. FSi/FSm).}

\item{input.obs.grid}{Optional. The input grid of observations passed to the plotting function.
Some metadata are extracted from this object (x-y axis ranges)}

\item{input.fcst.grid}{Optional. The input forecast grid passed to the plotting function.
Some metadata are extracted from this object (x-y axis ranges)}

\item{arg.list}{Argument list (A named list in key-value format).}

\item{FcstMetadata}{The metadata structure containing the step describing the forecast data provenance}

\item{ObsMetadata}{The metadata structure containing the step describing the observation data provenance}

\item{VerificationMetadata}{The metadata structure containing the step describing the verification provenance.
Only needed when EQC info (ACC) is added to the plume}
}
\description{
Build a directed metadata graph describing a ENSO plume plot
}
\details{
This function takes as reference the semantics defined in the Graphical Product ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\references{
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other graphical.products: \code{\link{metaclipR.AnomalyMap}},
  \code{\link{metaclipR.SkillMap}},
  \code{\link{metaclipR.SpatialPlot}}
}
\author{
D. San Martín, J. Bedia
}
