% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.loadeR.R
\name{metaclipR.loadeR}
\alias{metaclipR.loadeR}
\title{Directed metadata graph construction for loadeR outputs}
\usage{
metaclipR.loadeR(package = "loadeR", version = "1.4.0", graph, fun = NULL,
  arg.list = NULL, index.name = "none", output = NULL)
}
\arguments{
\item{package}{loading climate4R package. Default to \code{"loadeR"} (but see argument \code{fun})}

\item{version}{package version}

\item{graph}{A previously initialized metaclipR output list ($graph + $parentnodename).}

\item{fun}{function name. Default to \code{"loadGridData"}, but other loading functions are accepted, namely 
loadStationData from package \pkg{loadeR}, \code{loadECOMS}, from package \pkg{loadeR.ECOMS} or \code{loadValueStations}
 and \code{loadValuePredictions} from package \pkg{R.VALUE}.}

\item{arg.list}{arg.list. See details}

\item{index.name}{Xharacter string (default to "none"). In case of using loadGridData, it exists the possibility of
applying filters to monthly aggregations, leading to specific climate indices with particular names.
 Some of these are instances of the ds:ClimateIndex class (e.g. "SU", for the ETCCDI index 'summer days'). 
 An overview of already existing indices see \code{knownClassIndividuals("ClimateIndex")}.}

\item{output}{A climate4R grid, resulting from the application of the specified \code{fun} function. 
The name of the grid is passed as a character string}
}
\value{
A named list with the updated graph in element \code{graph} and the terminal node in \code{parentnodename},
needed for linking subsequent operations.

\strong{Argument list}

A list of arguments passed to \code{fun} is required. The different 
arguments are explained in the the help menu of the corresponding functions.
The following list need to be specified. The default values are also indicated:

\itemize{
\item \code{var = NULL}
\item \code{lonLim = NULL}
\item \code{latLim = NULL} 
\item \code{season = NULL}
\item \code{members = NULL} 
\item \code{years = NULL} 
}
}
\description{
Build a directed metadata graph from loadeR outputs, that are subsets (and possibly 
aggregations and/or climate indices) of Datasets.
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\references{
\href{http://http://metaclip.predictia.es/}{The METACLIP web page at Predictia}

\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other subsetting: \code{\link{metaclipR.DatasetSubset}}
}
\author{
D. San Martín, J. Bedia
}
