% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.RectangularGrid.R
\name{metaclipR.RectangularGrid}
\alias{metaclipR.RectangularGrid}
\title{Set a reference RectangularGrid}
\usage{
metaclipR.RectangularGrid(resX, resY, xmin = NULL, xmax = NULL,
  ymin = NULL, ymax = NULL, dc.description = NULL)
}
\arguments{
\item{resX}{Required. Spatial resolution of the grid in the X-coordinates}

\item{resY}{Required. Same as \code{resX}, but for the Y-coordinates}

\item{xmin}{Optional. The minimum x coordinate defining the spatial extent}

\item{xmax}{Same as \code{xmin}, but the maximum}

\item{ymin}{Same as \code{xmin}, but for the Y coordinates}

\item{ymax}{Same as \code{ymin}, but for the maximum Y coordinate.}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:RectangularGrid node.}
}
\value{
A metaclipR object (igraph-class structure + terminal node name)
}
\description{
Creates a graph containing a RectangularGrid definition, 
that can be used later as spatial reference for regridding operations etc. so all steps point to the same reference grid
}
\author{
J Bedia
}
