% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.SkillMap.R
\name{metaclipR.SkillMap}
\alias{metaclipR.SkillMap}
\title{Directed metadata graph construction for skill Map graphical products}
\usage{
metaclipR.SkillMap(graph, package = "visualizeR", version = "1.1.1",
  input.grid = NULL, fun = "skillMap", arg.list = NULL,
  StipplingInputGraph = NULL, epsg.code = "EPSG:4979")
}
\arguments{
\item{graph}{metaclipR output object containing the input grid whose values are to be mapped}

\item{package}{package}

\item{version}{version}

\item{input.grid}{Optional. The input grid passed to the plotting function. Some metadata are extracted from 
this object if not explicit from the plotting arguments (e.g., the spatial extent of the map).}

\item{fun}{function name. Default to \code{spatialPlot}, from package \code{visualizeR}.}

\item{arg.list}{Argument list (A named list in key-value format).}

\item{StipplingInputGraph}{Currently unused. The metaclipR output containing the step from which the stippling is derived (default to \code{NULL},
and no stippling is indicated)}

\item{epsg.code}{A character string indicating the EPSG code of the map projection. Default to \code{"4979"}
(the most widely used WGS84 proj).}
}
\description{
Build a directed metadata graph describing a skill map
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\references{
\href{https://docs.google.com/presentation/d/1Bn7M7IIFVvOoKG7YymXkDnNH4vJOqO6xmimeUI6Uxc0/view#slide=id.g23b883c69b_0_194}{Visual schema of the graphical product ontology}
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other graphical.products: \code{\link{metaclipR.AnomalyMap}},
  \code{\link{metaclipR.EnsoPlume}},
  \code{\link{metaclipR.SpatialPlot}}
}
\author{
D. San Martín, J. Bedia
}
