% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.SpatialExtent.R
\name{metaclipR.SpatialExtent}
\alias{metaclipR.SpatialExtent}
\title{Set a reference SpatialExtent}
\usage{
metaclipR.SpatialExtent(xyCoords, region = NULL)
}
\arguments{
\item{xyCoords}{Grid coordinates (i.e., the \code{$xyCoords} element of a \pkg{climate4R} object)}

\item{region}{An optional character string designating a known region. See \code{knownClassIndividuals("HorizontalExtent")}
for a list of regions defined in the current stable vocabulary version. Default to \code{NULL} and unused.}
}
\value{
A metaclipR list (igraph-class structure + terminal node name)
}
\description{
Creates a graph containing a SpatialExtent definition, that can be used later as spatial reference of
suubsetting, regridding operations etc. so all steps point to the same reference spatial extent
}
\author{
J Bedia
}
