% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.spatialPlot.R
\name{metaclipR.SpatialPlot}
\alias{metaclipR.SpatialPlot}
\title{Directed metadata graph construction for spatialPlot outputs}
\usage{
metaclipR.SpatialPlot(package = "visualizeR", version = "1.1.1", graph,
  input.grid = NULL, type = "grid", fun = "spatialPlot",
  arg.list = NULL, stipplingInputGraph = NULL, epsg.code = "EPSG:4329")
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{graph}{i-graph class object containing the input grid whose values are to be mapped}

\item{input.grid}{Optional. The input grid passed to the plotting function. Some metadata are extracted from 
this object if not explicit from the plotting arguments (e.g., the spatial extent of the map).}

\item{type}{The type of input grid passed to \code{spatialPlot}. Default to \code{"grid"}. The other possible choice is \code{"stations"},
for station datasets. Whie the first is mapped to the go:MapRaster class, the latter is assigned to go:MapPoints.}

\item{fun}{function name. Default to \code{spatialPlot}, from package \code{visualizeR}.}

\item{arg.list}{Argument list (A named list in key-value format).}

\item{stipplingInputGraph}{Currently unused. The graph containing the step from which the stippling is derived (default to \code{NULL},
and no stippling is indicated)}

\item{epsg.code}{A character string indicating the EPSG code of the map projection. Default to \code{"4979"}}
}
\description{
Build a directed metadata graph describing a spatialPlot in visualizeR
}
\details{
It is important to note that the resulting output of this function will reatin as parent node the last step before plotting,
 so intermediate plots within the data workflow can be recorded in the provenance track without affecting
 subsequent operations.

This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\references{
\href{http://www.meteo.unican.es/en/climate4r}{climate4R page at University of Cantabria}
}
\seealso{
Other graphical.products: \code{\link{metaclipR.AnomalyMap}},
  \code{\link{metaclipR.EnsoPlume}},
  \code{\link{metaclipR.SkillMap}}
}
\author{
D. San Martín, J. Bedia
}
