% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Validation.R
\name{metaclipR.Validation}
\alias{metaclipR.Validation}
\title{Directed metadata graph construction for validation steps}
\usage{
metaclipR.Validation(package, version, fun, QualityAspect, arg.list,
  PredictionGraph, ReferenceGraph)
}
\arguments{
\item{package}{Validation package.}

\item{version}{A character string of the package version (e.g. as returned by \code{\link[utils]{packageVersion}})}

\item{fun}{Validation function.}

\item{QualityAspect}{Class name. Quality Aspect addressed by the Validation. Possible values are 
\code{"Bias"}, \code{"Accuracy"}, \code{"Association"}, \code{"Reliability"}, \code{"Discrimination"} and \code{"Resolution"}.}

\item{arg.list}{argument list passed by \code{fun}.}

\item{PredictionGraph}{metaclipR output containing the Predictions/Projections to validate}

\item{ReferenceGraph}{metaclipR output containing the reference (observations)}
}
\description{
Build a directed metadata graph from validation routines
}
\details{
This function takes as reference the semantics defined in the Verification ontology defined in the 
Metaclip Framework (\url{http://metaclip.predictia.es/}).
}
\seealso{
Other validation: \code{\link{metaclipR.easyVerification}}
}
\author{
D. San Martín, J. Bedia
}
