% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{my_add_vertices}
\alias{my_add_vertices}
\title{Internal helper for vertex addition}
\usage{
my_add_vertices(graph, name = NULL, nv = 1, label = NULL,
  className = NULL, attr = NULL)
}
\arguments{
\item{graph}{i-graph class graph on which the node is added (might be an empty graph if it is the initial Dataset node)}

\item{name}{vertex name}

\item{nv}{Integer. Number of vertices added. Default (and only sensible value) to 1.}

\item{label}{A \dQuote{rdfs:label} data property, useful for displaying the graph in visualization APIs}

\item{className}{Character. Name of the class that the node blengs to.
Requires prefix inclusion (e.g. \code{className = "ds:DatasetSubset"})}

\item{description}{Optional. Description of the node (actually unused but for developer internal checks)}
}
\description{
A wrapper of \code{igraph::add_vertex} to avoid adding vertex refering to individuals already defined in the graph
}
\author{
J Bedia
}
\keyword{internal}
