% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{my_union}
\alias{my_union}
\title{Attribute-preserving igraph union}
\usage{
my_union(g, g1, g2)
}
\arguments{
\item{g}{The union graph}

\item{g1}{First input graph}

\item{g2}{Second input graph}
}
\value{
Invisible returns a new graph with fixed attributes
}
\description{
Transforms the resulting output from igraph::union to adequately recover the original attributes.
Apart from vertex attributes, the "label" attribute from edges, used to define object properties, is fixed.
}
\author{
J. Bedia
}
\keyword{internal}
