% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyJSON.R
\name{prettyJSON}
\alias{prettyJSON}
\title{Prettify or minify a METACLIP JSON}
\usage{
prettyJSON(jsonfile, indent = 2, minify = FALSE)
}
\arguments{
\item{jsonfile}{Path to the JSON file}

\item{indent}{Positive integer. Number of spaces for indentation (ignored when \code{minify = TRUE})}

\item{minify}{Logical. If set to \code{TRUE}, removes all indentation and whitespaces to minimize the file size.}
}
\value{
The input file is overwritten by its prettyfied/minified version
}
\description{
A wrapper of \code{\link[jsonlite]{prettify}} and \code{\link[jsonlite]{minify}} to modify METACLIP outputs for either readability or minimum size.
}
\seealso{
[graph2json()]
}
\author{
juaco
}
