% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{randomName}
\alias{randomName}
\title{Generate a random character string}
\usage{
randomName(len = 6)
}
\arguments{
\item{len}{length of the character string}
}
\value{
A string with the specified number of characters
}
\description{
Generates a random character string (lowercase) of a specified length. 
Used to avoid duplicated node names in igraph
}
\examples{
randomName()
randomName(10)
}
\author{
J Bedia
}
\keyword{internal}
