% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2json.R
\name{serializeVertex}
\alias{serializeVertex}
\title{recursive function for graph serialization}
\usage{
serializeVertex(g, vertex, describedNodeNames, connection)
}
\description{
Recursive function to trace and encode properties of nodes
}
\keyword{internal}
