% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{setNodeName}
\alias{setNodeName}
\title{Set a node name}
\usage{
setNodeName(node.name, node.class, vocabulary = "datasource")
}
\arguments{
\item{node.name}{Proposed node name. This ca be an individual instance}

\item{node.class}{Class of the node}

\item{vocabulary}{Vocabulary defining the \code{node.class}. Default to \code{"datasource"}.}
}
\value{
A character string to be set as the \code{"name"} attribute by \code{\link{my_add_vertices}}
}
\description{
Set a node name after checking whether it is an individual instance or not
}
