% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{showUDGDatasources}
\alias{showUDGDatasources}
\title{Table of User Data Gateway PUBLIC datasets}
\usage{
showUDGDatasources()
}
\value{
A \code{data.frame}
}
\description{
Show the internal table with known Datasets and relevant metadata
}
\examples{
showUDGDatasources()[c(2,35,60,81),]
}
\author{
J Bedia
}
