% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip-helpers.R
\name{getIndividualClass}
\alias{getIndividualClass}
\title{Identify the class belonging of an individual}
\usage{
getIndividualClass(individual.name, vocabulary = "datasource")
}
\arguments{
\item{individual.name}{Name of the individual}

\item{vocabulary}{The target vocabulary name. Possible values are \code{"datasource"} (the default),
\code{"calibration"}, \code{"verification"} and \code{"graphical_output"}.}
}
\value{
A character string with the most direct class belonging (superclasses are not indicated)
}
\description{
Identifies the class belonging of an individual. Superclasses are ignored
}
\details{
The function will check the existing individuals in the latest stable target 
ontology release
}
\note{
The function won't work if there is no internet connection, or any other connection problem prevents
 read access to the ontology file.
}
\examples{
getIndividualClass("UDG") # The default vocabulary is datasource (can be omitted)
getIndividualClass("EQM", vocabulary = "calibration")
}
\seealso{
Other ontology.helpers: \code{\link{knownClassIndividuals}}
}
\author{
J Bedia, D. San-Martín
}
\concept{ontology.helpers}
