% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclip.graph.Command.R
\name{metaclip.graph.Command}
\alias{metaclip.graph.Command}
\title{Directed metadata graph construction for Package/Command/Argument provenance description}
\usage{
metaclip.graph.Command(graph, package, version, fun, arg.list, origin.node.name)
}
\arguments{
\item{graph}{An igraph-class graph}

\item{package}{Character string. Package name}

\item{version}{Character string. Package version}

\item{fun}{Character string. Function name}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{origin.node.name}{Name of the origin node of the previously existing graph.}
}
\description{
Build a directed metadata graph describing R calls generating outputs
}
\details{
This function takes as reference the semantics defined in the Seasonal Forecast Verification ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
seems like a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

\seealso{
Other transformation: \code{\link{metaclipR.Aggregation}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Climatology}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Ensemble}},
  \code{\link{metaclipR.Interpolation}},
  \code{\link{metaclipR.etccdi}}
}
\author{
D. San Martín, J. Bedia
}
\keyword{internal}
