% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Aggregation.R
\name{metaclipR.Aggregation}
\alias{metaclipR.Aggregation}
\title{Directed metadata graph construction for Transformations of DatasetSubsets}
\usage{
metaclipR.Aggregation(package = "transformeR", version = "1.3.2", graph,
  fun = "aggregateGrid", arg.list = NULL)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{graph}{A previous metaclipR data structure from which the current step follows}

\item{fun}{function name. Unused (set to \code{"aggregateGrid"})}

\item{arg.list}{Argument list. See details}
}
\value{
A named list with the updated graph in element \code{"graph"} and the parent node name (\code{"parentnodename"}),
needed for linking subsequent operations.
}
\description{
Build a directed metadata graph describing a Transformation on a 
climate4R grid that is a subset of Dataset.
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).

\strong{Argument list}

The following list of arguments is required to define an aggregation:
\itemize{
\item \code{aggr.mem = list(FUN = NULL)}
\item \code{aggr.d = list(FUN = NULL)}
\item \code{aggr.m = list(FUN = NULL)}
\item \code{aggr.y = list(FUN = NULL)}
\item \code{aggr.lat = list(FUN = NULL)}
\item \code{weight.by.lat = TRUE}
\item \code{aggr.lon = list(FUN = NULL)}
}

The different arguments are explained in the the help page of \code{\link[transformeR]{aggregateGrid}}
}
\examples{
\dontrun{
require(transformeR)
require(igraph)
pkg <- "transformeR"
v <- "1.1.1"
# Assume a given hindcast DatasetSubset (we first simplify a)
data("CFS_Iberia_hus850")
DS <- subsetGrid(CFS_Iberia_hus850, members = 1:3, years = 1989:1991)
graph <- metaclipR.DatasetSubset(package = pkg,
                                 version = v,
                                 arg.list = list(members = 1:3,
                                                 years = 1989:1991),
                                 fun = "subsetGrid",
                                 output = "DS")
# An aggregation is performed on the example data. In this case, 
# The forecast is aggregated by members (ensemble mean), and in space
# (along longitude and latitude) to obtain a spatial mean. Original data is daily,
# and an annual aggregation is performed, considering the monthly means and the
# annual maxima:
fun <- "aggregateGrid"
arg.list <- list("aggr.mem" = list(FUN = "mean", na.rm = TRUE),
                 "aggr.d" = list(FUN = NULL),
                 "aggr.m" = list(FUN = "mean", na.rm = TRUE),
                 "aggr.y" = list(FUN = "max", na.rm = TRUE),
                 "aggr.lat" = list(FUN = "mean", na.rm = TRUE),
                 "aggr.lon" = list(FUN = "mean", na.rm = TRUE),
                 "weight.by.lat" = TRUE)
# The aggregation is undertaken with transformeR::subsetGrid:
arg.list[["grid"]] <- DS
out <- do.call(fun, arg.list)
# This is how metadata is encoded:
# 1.) We identify the origin node from which the first transformation hangs:
graph$parentnodename
# 2.) metaclipR.Aggregation is called:
graph <- metaclipR.Aggregation(package = pkg,
                               version = v,
                               graph = graph,
                               fun = fun,
                               arg.list = arg.list)
# And this is the metadata description stored in the igraph-class object:
plot(graph$graph)
}
}
\references{
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other transformation: \code{\link{metaclip.graph.Command}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Climatology}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Regridding}},
  \code{\link{metaclipR.etccdi}}
}
\author{
D. San Martín, J. Bedia
}
