% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Binding.R
\name{metaclipR.Binding}
\alias{metaclipR.Binding}
\title{Construct a METACLIP representation of the ds:Binding class}
\usage{
metaclipR.Binding(graph.list, dim.along, dc.description = NULL)
}
\arguments{
\item{graph.list}{A \code{metaclipR} object data list, each element being the subelements to be binded}

\item{dim.along}{Dimension along which data are binded. Accepted values are \code{"time"}, \code{"lat"} and \code{"lon"}. See Details.}

\item{dc.description}{Default to \code{NULL} and unused. Otherwise, this is a character string that will be appendend as a
"dc:description" annotation to the ds:Ensemble-class node.}
}
\description{
Build a directed metadata graph describing a binding transformation from two or more
ds:Steps along a chosen dimension
}
\details{
The special case in which different subsets of the same dataset are binded along the member dimension is sepparately treated by the Ensemble class,
described via \code{\link{metaclipR.Ensemble}}.

This function takes as reference the semantics defined in the Datasource and Transformation ontology (\url{www.metaclip.org/datasource/datasource.owl})
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\seealso{
Other transformation: \code{\link{metaclip.graph.Command}},
  \code{\link{metaclipR.Aggregation}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Climatology}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Ensemble}},
  \code{\link{metaclipR.Interpolation}},
  \code{\link{metaclipR.etccdi}}
}
\author{
J. Bedia
}
\concept{transformation}
