% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Climatology.R
\name{metaclipR.Climatology}
\alias{metaclipR.Climatology}
\title{Directed metadata graph construction for Climatology transformations}
\usage{
metaclipR.Climatology(package = "transformeR", version = "1.3.2", graph,
  fun = "climatology", arg.list = NULL)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{graph}{A previous metaclipR data structure from which the current step follows}

\item{fun}{function name. Unused (set to \code{"climatology"})}

\item{arg.list}{Argument list. See details.}
}
\value{
A named list with the updated graph in element \code{"graph"} and the parent node name (\code{"parentnodename"}),
needed for linking subsequent operations.
}
\description{
Build a directed metadata graph describing a Transformation on a 
climate4R grid that is a Climatology
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).

\strong{Argument list}

The following list of arguments is required to define an aggregation:
\itemize{
\item \code{clim.fun}
\item \code{by.member}
}

The climatology function will be encoded as the ds:hasCellMethod data property of the Climatology.
Further optional arguments can be included in orde to enrich the provenance information. 
The different arguments are explained in the the help page of \code{\link[transformeR]{climatology}}.

Note that if no argument list is provided at all, the default argument values (i.e. a mean cell method and by members),
 will be assumed.
}
\references{
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other transformation: \code{\link{metaclip.graph.Command}},
  \code{\link{metaclipR.Aggregation}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Ensemble}},
  \code{\link{metaclipR.Regridding}},
  \code{\link{metaclipR.etccdi}}
}
\author{
D. San Martín, J. Bedia
}
