% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Ensemble.R
\name{metaclipR.Ensemble}
\alias{metaclipR.Ensemble}
\title{Construct a METACLIP representation of the ds:Ensemble class}
\usage{
metaclipR.Ensemble(package = "transformeR", version = "1.3.3",
  output = NULL, fun = "bindGrid.member", arg.list = NULL, graph.list)
}
\arguments{
\item{package}{package name. Default to \code{"transformeR"}}

\item{version}{Package version string.}

\item{output}{Optional. The output R object name, as character string}

\item{fun}{function name. Default to \code{"bindGrid.member"})}

\item{arg.list}{Argument list. See details}

\item{graph.list}{A \code{metaclipR} object data list, each element being the
graph defining each ensemble member}
}
\description{
Build a directed metadata graph describing an Ensemble transformation from two or more
ds:Steps
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).

\strong{Argument list}

The following (minimal) list of arguments is required to define an ETCCDI climate index transformation:
\itemize{
\item \code{index.code}
}

Further optional arguments can be passed to \code{arg.list} for a more detailed description of the command call.
The different arguments are explained in the the help page of \code{\link[climate4R.climdex]{climdexGrid}}.
}
\references{
\href{https://github.com/Predictia/metaclip}{METACLIP Overview}

\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other transformation: \code{\link{metaclip.graph.Command}},
  \code{\link{metaclipR.Aggregation}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Climatology}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Regridding}},
  \code{\link{metaclipR.etccdi}}
}
\author{
D. San Martín, J. Bedia
}
