% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.EnsoPlume.R
\name{metaclipR.EnsoPlume}
\alias{metaclipR.EnsoPlume}
\title{Directed metadata graph construction for ENSO plume graphical products}
\usage{
metaclipR.EnsoPlume(package = "visualizeR", version = "1.1.1",
  fun = "ensoPlumeC3S", multimodel = FALSE, input.obs.grid = NULL,
  input.fcst.grid = NULL, arg.list = NULL, FcstMetadata,
  ObsMetadata = NULL, VerificationMetadata = NULL)
}
\arguments{
\item{package}{package}

\item{version}{version}

\item{fun}{function name.}

\item{multimodel}{Logical. Is the forecast done with a multimodel ensemble dataset?. Default
to \code{FALSE} (This is needed to ascertain the subtypes of ForecastSystemUncertainty communicated
by the product, i.e. FSi/FSm).}

\item{input.obs.grid}{Optional. The input grid of observations passed to the plotting function.
Some metadata are extracted from this object (x-y axis ranges)}

\item{input.fcst.grid}{Optional. The input forecast grid passed to the plotting function.
Some metadata are extracted from this object (x-y axis ranges)}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{FcstMetadata}{The metadata structure containing the step describing the forecast data provenance}

\item{ObsMetadata}{The metadata structure containing the step describing the observation data provenance}

\item{VerificationMetadata}{The metadata structure containing the step describing the verification provenance.
Only needed when EQC info (ACC) is added to the plume}
}
\description{
Build a directed metadata graph describing a ENSO plume plot
}
\details{
This function takes as reference the semantics defined in the Graphical Product ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org}).
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
seems like a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

\seealso{
Other graphical.products: \code{\link{metaclipR.AnomalyMap}},
  \code{\link{metaclipR.SkillMap}},
  \code{\link{metaclipR.SpatialPlot}}
}
\author{
D. San Martín, J. Bedia
}
\concept{graphical.products}
