% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.etccdi.R
\name{metaclipR.etccdi}
\alias{metaclipR.etccdi}
\title{Directed metadata graph construction for ETCDDI climate index transformations}
\usage{
metaclipR.etccdi(graph, package = "climate4R.climdex", version = "0.1.2",
  output = NULL, fun = "climdexGrid", arg.list = NULL)
}
\arguments{
\item{graph}{Output from previous metaclipR function. A list with an i-graph class object plus the name of the parent node 
from which the climate index step hangs.
\code{transformeR} indicated in argument \code{fun}.}

\item{package}{package}

\item{version}{version}

\item{output}{Optional. The output R object name, as character string}

\item{fun}{function name. Unused (set to \code{"climdexGrid"})}

\item{arg.list}{Argument list. See details}
}
\description{
Build a directed metadata graph describing a ETCDDI climate index transformation on a 
climate4R grid
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).

\strong{Argument list}

The following (minimal) list of arguments is required to define an ETCCDI climate index transformation:
\itemize{
\item \code{index.code}
}

Further optional arguments can be passed to \code{arg.list} for a more detailed description of the command call.
The different arguments are explained in the the help page of \code{\link[climate4R.climdex]{climdexGrid}}.
}
\examples{
\dontrun{
require(climate4R.climdex)
require(igraph)
data("tasmax.eobs")
a <- metaclipR.DatasetSubset(output = "tasmax.eobs",
                             fun = "subsetGrid",
                             arg.list = list(lonLim = c(-10,4.5),
                             latLim = c(35,44),
                             season = 1:12,
                             years = 1991:2010))

tx10p <- climdexGrid(index.code = "TX10p",
                     tx = tasmax.eobs,
                     index.arg.list = list(freq = "annual"))

arg.list <- list(index.code = "TX10p",
                 index.arg.list = list(freq = "annual"))

require(visualizeR)
spatialPlot(climatology(tx10p),
            backdrop.theme = "countries",
            main = "Mean percentage of days when TX < 10 degC (1991-2010)")
metadata <- metaclipR.etccdi(graph = a, arg.list = arg.list)            
plot(metadata$graph)
}
}
\references{
\href{https://github.com/Predictia/metaclip}{METACLIP Overview}

\href{http://visualdataweb.de/webvowl/#iri=http://metaclip.predictia.es/datasource/datasource.owl}{Visual schema of the data transformation ontology} 

\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other transformation: \code{\link{metaclip.graph.Command}},
  \code{\link{metaclipR.Aggregation}},
  \code{\link{metaclipR.AnomalyCalculation}},
  \code{\link{metaclipR.Climatology}},
  \code{\link{metaclipR.Dataset}},
  \code{\link{metaclipR.Ensemble}},
  \code{\link{metaclipR.Regridding}}
}
\author{
D. San Martín, J. Bedia
}
