% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.Interpolation.R
\name{metaclipR.Regridding}
\alias{metaclipR.Regridding}
\title{Directed metadata graph construction for Transformations of DatasetSubsets}
\usage{
metaclipR.Regridding(
  graph,
  package = "transformeR",
  version = "1.3.2",
  fun = "interpGrid",
  arg.list = NULL
)
}
\arguments{
\item{graph}{A previous metaclipR data structure from which the current step follows}

\item{package}{package}

\item{version}{version}

\item{fun}{function name. Unused (set to \code{"interpGrid"})}

\item{arg.list}{Argument list. See details}
}
\description{
Build a directed metadata graph describing a regridding Transformation on a 
climate4R grid
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://metaclip.predictia.es/}).

\strong{Argument list}

The following list of arguments is required to define an aggregation:
\itemize{
\item \code{new.coordinates}
\item \code{method}
\item \code{bilin.method}
}

The different arguments are explained in the the help page of \code{\link[transformeR]{interpGrid}}.
}
\references{
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
\url{metaclip.predictia.es}
}
\author{
D. San Martín, J. Bedia
}
