% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclipR.SkillMap.R
\name{metaclipR.SkillMap}
\alias{metaclipR.SkillMap}
\title{Directed metadata graph construction for skill Map graphical products}
\usage{
metaclipR.SkillMap(graph, package = "visualizeR", version = "1.1.1",
  input.grid = NULL, fun = "skillMap", arg.list = NULL,
  StipplingInputGraph = NULL, epsg.code = "EPSG:4979")
}
\arguments{
\item{graph}{metaclipR output object containing the input grid whose values are to be mapped}

\item{package}{package}

\item{version}{version}

\item{input.grid}{Optional. The input grid passed to the plotting function. Some metadata are extracted from 
this object if not explicit from the plotting arguments (e.g., the spatial extent of the map).}

\item{fun}{function name. Default to \code{spatialPlot}, from package \code{visualizeR}.}

\item{arg.list}{Either a character string containing a literal command call or a key-value pair list of arguments-values. See the dedicated section more for details.}

\item{StipplingInputGraph}{Currently unused. The metaclipR output containing the step from which the stippling is derived (default to \code{NULL},
and no stippling is indicated)}

\item{epsg.code}{A character string indicating the EPSG code of the map projection. Default to \code{"4979"}
(the most widely used WGS84 proj).}
}
\description{
Build a directed metadata graph describing a skill map
}
\details{
This function takes as reference the semantics defined in the Data Source and Transformation ontology
defined in the Metaclip Framework (\url{http://www.metaclip.org/}).
}
\section{About argument lists}{

Argument lists are key-value list in which the different arguments and their valus are passed 
to the command description. These were initially conceived as an adequate means of describing
command calls to interpreted languajes (R, python...) using both the Argument and Argument 
classes from the datasource vocabulary of METACLIP.

Alternatively, and in order to accomodate the description of the source code to other environments
and languajes (e.g. shell scripts, calls to CDOs etc.), the use of literal command calls 
seems like a more convenient choice. Literal command calls also allow for a more straightforward
reproducibility by just "copying and pasting" the code used to generate each step. Literal
command calls are encoded in METACLIP using the \emph{hadLiteralCommandCall} data property
from the datasource vocabulary, that is attached to the command description. 
To apply this second alternative, the \code{arg.list} argument is specified as a character string
containing the command call.

Note that default argument values are often omitted from comand calls. However, the default behaviour of 
commands may change depending on the specific software version, and it is therefore often advisable to 
explicitly include the default argument values for a more straightforward interpretation of the source code
by the user.
}

\references{
\href{https://docs.google.com/presentation/d/1Bn7M7IIFVvOoKG7YymXkDnNH4vJOqO6xmimeUI6Uxc0/view#slide=id.g23b883c69b_0_194}{Visual schema of the graphical product ontology}
\href{http://www.meteo.unican.es/en/climate4r}{Climate4R page at University of Cantabria}
}
\seealso{
Other graphical.products: \code{\link{metaclipR.AnomalyMap}},
  \code{\link{metaclipR.EnsoPlume}},
  \code{\link{metaclipR.SpatialPlot}}
}
\author{
D. San Martín, J. Bedia
}
